62 Wantage Close
"'My name,' said he, 'is Mr. Duncan Ross, and I am myself one of the pensioners upon the fund left
by our noble benefactor. Are you a married man, Mr. Wilson? Have you a family?'

"I answered that I had not.

"His face fell immediately.

"'Dear me!' he said gravely, 'that is very serious indeed! I am sorry to hear you say that. The fund
was, of course, for the propagation and spread of the red-heads as well as for their maintenance. It
is exceedingly unfortunate that you should be a bachelor.'

"My face lengthened at this, Mr. Holmes, for I thought that I was not to have the vacancy after all;
but after thinking it over for a few minutes he said that it would be all right.

"'In the case of another,' said he, 'the objection might be fatal, but we must stretch a point in
favour of a man with such a head of hair as yours. When shall you be able to enter upon your new
duties?'