127 Moor End
"There can be no question as to their nature," he answered.

"Then what are they? Who is this K. K. K., and why does he pursue this unhappy family?"

Sherlock Holmes closed his eyes and placed his elbows upon the arms of his chair, with his finger-
tips together. "The ideal reasoner," he remarked, "would, when he had once been shown a single fact
in all its bearings, deduce from it not only all the chain of events which led up to it but also all
the results which would follow from it. As Cuvier could correctly describe a whole animal by the
contemplation of a single bone, so the observer who has thoroughly understood one link in a series
of incidents should be able to accurately state all the other ones, both before and after. We have
not yet grasped the results which the reason alone can attain to. Problems may be solved in the
study which have baffled all those who have sought a solution by the aid of their senses. To carry
the art, however, to its highest pitch, it is necessary that the reasoner should be able to utilise
all the facts which have come to his knowledge; and this in itself implies, as you will readily see,
a possession of all knowledge, which, even in these days of free education and encyclopaedias, is a
somewhat rare accomplishment. It is not so impossible, however, that a man should possess all
knowledge which is likely to be useful to him in his work, and this I have endeavoured in my case to
do. If I remember rightly, you on one occasion, in the early days of our friendship, defined my
limits in a very precise fashion."

"Yes," I answered, laughing. "It was a singular document. Philosophy, astronomy, and politics were
marked at zero, I remember. Botany variable, geology profound as regards the mud-stains from any
region within fifty miles of town, chemistry eccentric, anatomy unsystematic, sensational literature
and crime records unique, violin-player, boxer, swordsman, lawyer, and self-poisoner by cocaine and
tobacco. Those, I think, were the main points of my analysis."

Holmes grinned at the last item. "Well," he said, "I say now, as I said then, that a man should keep
his little brain-attic stocked with all the furniture that he is likely to use, and the rest he can
put away in the lumber-room of his library, where he can get it if he wants it. Now, for such a case
as the one which has been submitted to us to-night, we need certainly to muster all our resources.
Kindly hand me down the letter K of the 'American Encyclopaedia' which stands upon the shelf beside
you. Thank you. Now let us consider the situation and see what may be deduced from it. In the first
place, we may start with a strong presumption that Colonel Openshaw had some very strong reason for
leaving America. Men at his time of life do not change all their habits and exchange willingly the
charming climate of Florida for the lonely life of an English provincial town. His extreme love of
solitude in England suggests the idea that he was in fear of someone or something, so we may assume
as a working hypothesis that it was fear of someone or something which drove him from America. As to
what it was he feared, we can only deduce that by considering the formidable letters which were
received by himself and his successors. Did you remark the postmarks of those letters?"