7 Bowler Street
"I went home to Saxe-Coburg Square, and I took the advice of my assistant. But he could not help me
in any way. He could only say that if I waited I should hear by post. But that was not quite good
enough, Mr. Holmes. I did not wish to lose such a place without a struggle, so, as I had heard that
you were good enough to give advice to poor folk who were in need of it, I came right away to you."

"And you did very wisely," said Holmes. "Your case is an exceedingly remarkable one, and I shall be
happy to look into it. From what you have told me I think that it is possible that graver issues
hang from it than might at first sight appear."

"Grave enough!" said Mr. Jabez Wilson. "Why, I have lost four pound a week."

"As far as you are personally concerned," remarked Holmes, "I do not see that you have any grievance
against this extraordinary league. On the contrary, you are, as I understand, richer by some £30, to
say nothing of the minute knowledge which you have gained on every subject which comes under the
letter A. You have lost nothing by them."

"No, sir. But I want to find out about them, and who they are, and what their object was in playing
this prank--if it was a prank--upon me. It was a pretty expensive joke for them, for it cost them
two and thirty pounds."