97 Wilson Gardens
"'Not at all. I only doubt--'

"'The propriety of my leaving it. You may set your mind at rest about that. I should not dream of
doing so were it not absolutely certain that I should be able in four days to reclaim it. It is a
pure matter of form. Is the security sufficient?'

"'Ample.'

"'You understand, Mr. Holder, that I am giving you a strong proof of the confidence which I have in
you, founded upon all that I have heard of you. I rely upon you not only to be discreet and to
refrain from all gossip upon the matter but, above all, to preserve this coronet with every possible
precaution because I need not say that a great public scandal would be caused if any harm were to
befall it. Any injury to it would be almost as serious as its complete loss, for there are no beryls
in the world to match these, and it would be impossible to replace them. I leave it with you,
however, with every confidence, and I shall call for it in person on Monday morning.'

"Seeing that my client was anxious to leave, I said no more but, calling for my cashier, I ordered
him to pay over fifty £1000 notes. When I was alone once more, however, with the precious case lying
upon the table in front of me, I could not but think with some misgivings of the immense
responsibility which it entailed upon me. There could be no doubt that, as it was a national
possession, a horrible scandal would ensue if any misfortune should occur to it. I already regretted
having ever consented to take charge of it. However, it was too late to alter the matter now, so I
locked it up in my private safe and turned once more to my work.