52 Balcombe Close
"Experience," said Holmes, laughing. "Indirectly it may be of value, you know; you have only to put
it into words to gain the reputation of being excellent company for the remainder of your
existence."

The Lord St. Simon marriage, and its curious termination, have long ceased to be a subject of
interest in those exalted circles in which the unfortunate bridegroom moves. Fresh scandals have
eclipsed it, and their more piquant details have drawn the gossips away from this four-year-old
drama. As I have reason to believe, however, that the full facts have never been revealed to the
general public, and as my friend Sherlock Holmes had a considerable share in clearing the matter up,
I feel that no memoir of him would be complete without some little sketch of this remarkable
episode.