139 Glendale
"The facts are briefly these: Some five years ago, during a lengthy visit to Warsaw, I made the
acquaintance of the well-known adventuress, Irene Adler. The name is no doubt familiar to you."

"Kindly look her up in my index, Doctor," murmured Holmes without opening his eyes. For many years
he had adopted a system of docketing all paragraphs concerning men and things, so that it was
difficult to name a subject or a person on which he could not at once furnish information. In this
case I found her biography sandwiched in between that of a Hebrew rabbi and that of a staff-
commander who had written a monograph upon the deep-sea fishes.

"Let me see!" said Holmes. "Hum! Born in New Jersey in the year 1858. Contralto--hum! La Scala, hum!
Prima donna Imperial Opera of Warsaw--yes! Retired from operatic stage--ha! Living in London--quite
so! Your Majesty, as I understand, became entangled with this young person, wrote her some
compromising letters, and is now desirous of getting those letters back."

"Precisely so. But how--"

"Was there a secret marriage?"

"None."