88 Wantage Close
"Oh, yes! he is the green-grocer who brings our vegetables round.  His name is Francis Prosper."

"He stood," said Holmes, "to the left of the door--that is to say, farther up the path than is
necessary to reach the door?"

"Yes, he did."

"And he is a man with a wooden leg?"

Something like fear sprang up in the young lady's expressive black eyes. "Why, you are like a
magician," said she. "How do you know that?" She smiled, but there was no answering smile in Holmes'
thin, eager face.