139 Holmefield Avenue
"There is some building going on there," said Holmes; "that is where we are going."

"There's the village," said the driver, pointing to a cluster of roofs some distance to the left;
"but if you want to get to the house, you'll find it shorter to get over this stile, and so by the
foot-path over the fields. There it is, where the lady is walking."

"And the lady, I fancy, is Miss Stoner," observed Holmes, shading his eyes. "Yes, I think we had
better do as you suggest."

We got off, paid our fare, and the trap rattled back on its way to Leatherhead.

"I thought it as well," said Holmes as we climbed the stile, "that this fellow should think we had
come here as architects, or on some definite business. It may stop his gossip. Good-afternoon, Miss
Stoner. You see that we have been as good as our word."

Our client of the morning had hurried forward to meet us with a face which spoke her joy. "I have
been waiting so eagerly for you," she cried, shaking hands with us warmly. "All has turned out
splendidly. Dr. Roylott has gone to town, and it is unlikely that he will be back before evening."