94 Badgers Dene
"I assure you, Watson, without affectation, that the status of my client is a matter of less moment
to me than the interest of his case. It is just possible, however, that that also may not be wanting
in this new investigation. You have been reading the papers diligently of late, have you not?"

"It looks like it," said I ruefully, pointing to a huge bundle in the corner. "I have had nothing
else to do."

"It is fortunate, for you will perhaps be able to post me up. I read nothing except the criminal
news and the agony column. The latter is always instructive. But if you have followed recent events
so closely you must have read about Lord St. Simon and his wedding?"

"Oh, yes, with the deepest interest."

"That is well. The letter which I hold in my hand is from Lord St. Simon. I will read it to you, and
in return you must turn over these papers and let me have whatever bears upon the matter. This is
what he says:

"'My dear Mr. Sherlock Holmes: "'Lord Backwater tells me that I may place implicit reliance upon
your judgment and discretion. I have determined, therefore, to call upon you and to consult you in
reference to the very painful event which has occurred in connection with my wedding. Mr. Lestrade,
of Scotland Yard, is acting already in the matter, but he assures me that he sees no objection to
your co-operation, and that he even thinks that it might be of some assistance. I will call at four
o'clock in the afternoon, and, should you have any other engagement at that time, I hope that you
will postpone it, as this matter is of paramount importance. "'Yours faithfully, "'St. Simon.'