125 Badgers Dene
"'But,' said I, 'there would be millions of red-headed men who would apply.'

"'Not so many as you might think,' he answered. 'You see it is really confined to Londoners, and to
grown men. This American had started from London when he was young, and he wanted to do the old town
a good turn. Then, again, I have heard it is no use your applying if your hair is light red, or dark
red, or anything but real bright, blazing, fiery red. Now, if you cared to apply, Mr. Wilson, you
would just walk in; but perhaps it would hardly be worth your while to put yourself out of the way
for the sake of a few hundred pounds.'