2 Badgers Dene
"It seemed to me that I had never met so fascinating and so thoughtful a man. As I was already in
debt to my tradesmen, the advance was a great convenience, and yet there was something unnatural
about the whole transaction which made me wish to know a little more before I quite committed
myself.

"'May I ask where you live, sir?' said I.

"'Hampshire. Charming rural place. The Copper Beeches, five miles on the far side of Winchester. It
is the most lovely country, my dear young lady, and the dearest old country-house.'

"'And my duties, sir? I should be glad to know what they would be.'