72 Harvey Street
"And that was why she shrieked and fainted when she saw the coronet," cried Mr. Holder. "Oh, my God!
what a blind fool I have been! And his asking to be allowed to go out for five minutes! The dear
fellow wanted to see if the missing piece were at the scene of the struggle. How cruelly I have
misjudged him!"

"When I arrived at the house," continued Holmes, "I at once went very carefully round it to observe
if there were any traces in the snow which might help me. I knew that none had fallen since the
evening before, and also that there had been a strong frost to preserve impressions. I passed along
the tradesmen's path, but found it all trampled down and indistinguishable. Just beyond it, however,
at the far side of the kitchen door, a woman had stood and talked with a man, whose round
impressions on one side showed that he had a wooden leg. I could even tell that they had been
disturbed, for the woman had run back swiftly to the door, as was shown by the deep toe and light
heel marks, while Wooden-leg had waited a little, and then had gone away. I thought at the time that
this might be the maid and her sweetheart, of whom you had already spoken to me, and inquiry showed
it was so. I passed round the garden without seeing anything more than random tracks, which I took
to be the police; but when I got into the stable lane a very long and complex story was written in
the snow in front of me.