22 Holmefield Avenue
"He shot one of his suspicious looks at me. 'Perhaps we had better proceed to business, then,' said
he. 'Mr. Ferguson and I will take you up to see the machine.'

"'I had better put my hat on, I suppose.'

"'Oh, no, it is in the house.'

"'What, you dig fuller's-earth in the house?'

"'No, no. This is only where we compress it. But never mind that. All we wish you to do is to
examine the machine and to let us know what is wrong with it.'