74 Cliff Place
"'This is my secretary and manager,' said the colonel. 'By the way, I was under the impression that
I left this door shut just now. I fear that you have felt the draught.'

"'On the contrary,' said I, 'I opened the door myself because I felt the room to be a little close.'