102 Balcombe Close
The station-master laughed heartily. "No, sir, Dr. Becher is an Englishman, and there isn't a man in
the parish who has a better-lined waistcoat. But he has a gentleman staying with him, a patient, as
I understand, who is a foreigner, and he looks as if a little good Berkshire beef would do him no
harm."

The station-master had not finished his speech before we were all hastening in the direction of the
fire. The road topped a low hill, and there was a great widespread whitewashed building in front of
us, spouting fire at every chink and window, while in the garden in front three fire-engines were
vainly striving to keep the flames under.

"That's it!" cried Hatherley, in intense excitement. "There is the gravel-drive, and there are the
rose-bushes where I lay. That second window is the one that I jumped from."