161 Down Close
"Not invisible but unnoticed, Watson. You did not know where to look, and so you missed all that was
important. I can never bring you to realise the importance of sleeves, the suggestiveness of thumb-
nails, or the great issues that may hang from a boot-lace. Now, what did you gather from that
woman's appearance? Describe it."

"Well, she had a slate-coloured, broad-brimmed straw hat, with a feather of a brickish red. Her
jacket was black, with black beads sewn upon it, and a fringe of little black jet ornaments. Her
dress was brown, rather darker than coffee colour, with a little purple plush at the neck and
sleeves. Her gloves were greyish and were worn through at the right forefinger. Her boots I didn't
observe. She had small round, hanging gold earrings, and a general air of being fairly well-to-do in
a vulgar, comfortable, easy-going way."

Sherlock Holmes clapped his hands softly together and chuckled.

"'Pon my word, Watson, you are coming along wonderfully. You have really done very well indeed. It
is true that you have missed everything of importance, but you have hit upon the method, and you
have a quick eye for colour. Never trust to general impressions, my boy, but concentrate yourself
upon details. My first glance is always at a woman's sleeve. In a man it is perhaps better first to
take the knee of the trouser. As you observe, this woman had plush upon her sleeves, which is a most
useful material for showing traces. The double line a little above the wrist, where the typewritist
presses against the table, was beautifully defined. The sewing-machine, of the hand type, leaves a
similar mark, but only on the left arm, and on the side of it farthest from the thumb, instead of
being right across the broadest part, as this was. I then glanced at her face, and, observing the
dint of a pince-nez at either side of her nose, I ventured a remark upon short sight and
typewriting, which seemed to surprise her."

"It surprised me."