78 Glendale
"I am afraid that you will find it difficult."

"Are you, indeed, now?" cried Lestrade with some bitterness. "I am afraid, Holmes, that you are not
very practical with your deductions and your inferences. You have made two blunders in as many
minutes. This dress does implicate Miss Flora Millar."

"And how?"

"In the dress is a pocket. In the pocket is a card-case. In the card-case is a note. And here is the
very note." He slapped it down upon the table in front of him. "Listen to this:

"'You will see me when all is ready. Come at once. "'F.H.M.'

Now my theory all along has been that Lady St. Simon was decoyed away by Flora Millar, and that she,
with confederates, no doubt, was responsible for her disappearance. Here, signed with her initials,
is the very note which was no doubt quietly slipped into her hand at the door and which lured her
within their reach."