97 Wantage Close
"That was what the county coroner asked me at the inquiry. It is my strong impression that I heard
it, and yet, among the crash of the gale and the creaking of an old house, I may possibly have been
deceived."

"Was your sister dressed?"

"No, she was in her night-dress. In her right hand was found the charred stump of a match, and in
her left a match-box."