3 Ventnor Terrace
"By the select prices. Eight shillings for a bed and eightpence for a glass of sherry pointed to one
of the most expensive hotels. There are not many in London which charge at that rate. In the second
one which I visited in Northumberland Avenue, I learned by an inspection of the book that Francis H.
Moulton, an American gentleman, had left only the day before, and on looking over the entries
against him, I came upon the very items which I had seen in the duplicate bill. His letters were to
be forwarded to 226 Gordon Square; so thither I travelled, and being fortunate enough to find the
loving couple at home, I ventured to give them some paternal advice and to point out to them that it
would be better in every way that they should make their position a little clearer both to the
general public and to Lord St. Simon in particular. I invited them to meet him here, and, as you
see, I made him keep the appointment."

"But with no very good result," I remarked. "His conduct was certainly not very gracious."

"Ah, Watson," said Holmes, smiling, "perhaps you would not be very gracious either, if, after all
the trouble of wooing and wedding, you found yourself deprived in an instant of wife and of fortune.
I think that we may judge Lord St. Simon very mercifully and thank our stars that we are never
likely to find ourselves in the same position. Draw your chair up and hand me my violin, for the
only problem we have still to solve is how to while away these bleak autumnal evenings."

"Holmes," said I as I stood one morning in our bow-window looking down the street, "here is a madman
coming along. It seems rather sad that his relatives should allow him to come out alone."